requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
pyngrok==7.0.1
gunicorn==21.2.0
python-json-logger==2.0.7
